% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kme_shrinkage_loocv.R
\name{A_lambda}
\alias{A_lambda}
\title{Calculates the A matrix}
\usage{
A_lambda(gram, regularized_inv)
}
\arguments{
\item{gram}{Matrix representing the Gram matrix.}

\item{regularized_inv}{The regularized inverse of the Gram matrix.}
}
\value{
The A matrix used in shrinkage estimation.
}
\description{
Calculates the A matrix
}
