% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{append_to_excel}
\alias{append_to_excel}
\title{Append Data to Excel File}
\usage{
append_to_excel(file_name, data)
}
\arguments{
\item{file_name}{A string specifying the name of the Excel file to append data to.}

\item{data}{A data frame containing the new simulation data to be appended.}
}
\value{
None
}
\description{
This function appends new simulation data to an existing Excel file.
}
\examples{
new_data <- data.frame(n = 10, lambda = 0.5, sigma = 0.1, error_rate = 0.02)
append_to_excel("simulation_results.xlsx", new_data)
}
