% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{aux_gibbs_exp_term}
\alias{aux_gibbs_exp_term}
\title{Calculate the exponential term in the Gibbs kernel function.}
\usage{
aux_gibbs_exp_term(a, b, length_scale = 1)
}
\arguments{
\item{a}{Numeric element of first vector}

\item{b}{Numeric element of second vector}

\item{length_scale}{Numeric scaling factor (default is 1)}
}
\value{
Numeric value representing the exponential term
}
\description{
Calculate the exponential term in the Gibbs kernel function.
}
\author{
Pouya Roudaki \email{roudaki.pouya@gmail.com}.
}
