% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R, R/kme_shrinkage.R
\name{aux_gibbs_kernel}
\alias{aux_gibbs_kernel}
\title{Calculate the Gibbs kernel between two vectors.}
\usage{
aux_gibbs_kernel(a, b, length_scale = 1, exp_term)

aux_gibbs_kernel(a, b, length_scale = 1, exp_term)
}
\arguments{
\item{a}{Numeric element of first vector}

\item{b}{Numeric element of second vector}

\item{length_scale}{Numeric scaling factor (default is 1)}

\item{exp_term}{Numeric value representing the exponential term}
}
\value{
Numeric value representing the Gibbs kernel

Numeric value representing the Gibbs kernel
}
\description{
Calculate the Gibbs kernel between two vectors.

Calculate the Gibbs kernel between two vectors.
}
\author{
Pouya Roudaki \email{roudaki.pouya@gmail.com}.
}
