% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{aux_gibbs_length}
\alias{aux_gibbs_length}
\title{Calculate the length of a vector after scaling.}
\usage{
aux_gibbs_length(a, length_scale = 1)
}
\arguments{
\item{a}{Numeric element of a vector}

\item{length_scale}{General numeric scaling factor (default is 1)}
}
\value{
Numeric scaling factor
}
\description{
Calculate the length of a vector after scaling.
}
\author{
Pouya Roudaki \email{roudaki.pouya@gmail.com}.
}
