% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kme_shrinkage_loocv.R
\name{B_lambda}
\alias{B_lambda}
\title{Calculates the B matrix}
\usage{
B_lambda(gram, regularized_inv)
}
\arguments{
\item{gram}{Matrix representing the Gram matrix.}

\item{regularized_inv}{The regularized inverse of the Gram matrix.}
}
\value{
The B matrix used in shrinkage estimation.
}
\description{
Calculates the B matrix
}
