% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kme_bayesian.R
\name{bayesian_est_KME}
\alias{bayesian_est_KME}
\title{Regularized Estimator of Kernel Mean Embedding(RKMSE)}
\usage{
bayesian_est_KME(
  evaluate_at,
  list_fixed,
  kernel_type = "rbf",
  kernel_params = list(length_scale = 1, degree = 2, free_add = 0, free_mult = 1,
    nu_matern = 1),
  dom_measure_type = "gaussian",
  dom_measure_parmas = list(eta = 1),
  tau = 1,
  precomputed = list(R = NULL, inverse_regularizer = NULL)
)
}
\arguments{
\item{evaluate_at}{The point at which to evaluate the kernel mean embedding estimator function.}

\item{list_fixed}{A list of fixed points for the kernel mean embedding estimator function.}

\item{kernel_type}{Type of kernel function (default is "rbf" for Radial Basis Function).}

\item{kernel_params}{A list of kernel parameters, including length_scale, degree, free_add, free_mult, and nu_matern.}
}
\value{
The regularized estimator of Kernel Mean Embedding.
}
\description{
Compute the regularized estimator of Kernel Mean Embedding. Same shrinkage in all directions
}
\examples{
bayesian_est_KME(evaluate_at = c(5, 3, 8), list_fixed = list(c(0, 1, 2), c(1, 2, 3)),
kernel_type = "rbf",
kernel_params = list(length_scale = 1, degree = 2, free_add = 0, free_mult = 1, nu_matern = 1),
dom_measure_type = "gaussian",
dom_measure_parmas = list(eta = 1),
tau = 1)
}
\references{
Bayesian Learning of Kernel Embeddings,
Association for Uncertainty in Artificial Intelligence
Equations (7), (8) and Appendix (A.3)
}
\author{
Pouya Roudaki \email{roudaki.pouya@gmail.com}.
}
