% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{beta_calculator}
\alias{beta_calculator}
\title{Calculate Betas for Different Estimator Types}
\usage{
beta_calculator(gram, folds, lambda, estimator_type)
}
\arguments{
\item{gram}{A square symmetric matrix representing the Gram matrix of the data.}

\item{folds}{A list of folds, each containing a train and test dataset with ids.}

\item{lambda}{A regularization parameter used for regularized and shrinkage estimators.}

\item{estimator_type}{A character string indicating the type of estimator: "standard",
"regularized", or "shrinkage".}
}
\value{
A list of calculated betas for each fold.
}
\description{
This function calculates the coefficients (betas) for each fold in a cross-validation
setup, based on the specified estimator type. It supports standard, regularized,
and shrinkage estimators by applying operations over each fold using `lapply`.
}
