% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife_grid.R
\name{calc_jackknife_error_nested_parallel}
\alias{calc_jackknife_error_nested_parallel}
\title{Parallelized Calculation of Jackknife Error}
\usage{
calc_jackknife_error_nested_parallel(
  lambda_hat,
  tau_hat,
  centered_kernel_mat_at_sampled,
  centered_kernel_mat_at_grid,
  centered_kernel_self_grid,
  sampled_x,
  x_grid,
  type_of_p_is_prob,
  type_of_q_is_prob,
  method_of_p_calculation,
  outer_index
)
}
\arguments{
\item{lambda_hat}{Estimated lambda parameter.}

\item{tau_hat}{Estimated tau parameter.}

\item{centered_kernel_mat_at_sampled}{A matrix of kernel values for the sampled points, centered.}

\item{centered_kernel_mat_at_grid}{A matrix of kernel values between sampled points and grid points, centered.}

\item{centered_kernel_self_grid}{A vector of self-kernel values for grid points.}

\item{sampled_x}{A vector of sampled input values.}

\item{x_grid}{A vector of grid points where calculations are performed.}

\item{type_of_p_is_prob}{Logical, whether the "p" kernel is interpreted as a probability.}

\item{type_of_q_is_prob}{Logical, whether the "q" kernel is interpreted as a probability.}

\item{method_of_p_calculation}{A string indicating the method to compute "p".}
}
\value{
The computed jackknife error as a numeric value.
}
\description{
This function computes the jackknife error for a given set of inputs, with parallelization
applied to the leave-one-out computations to improve performance.
}
\examples{
# Example usage of calc_jackknife_error_nested_parallel
jackknife_error <- calc_jackknife_error_nested_parallel(lambda_hat, tau_hat, kernel_sampled, kernel_grid,
                                                 kernel_self_grid, sampled_x, x_grid, TRUE, TRUE, "method")
}
