% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centered_kernel.R
\name{centered_kernel_matrix}
\alias{centered_kernel_matrix}
\title{Centered Kernel Matrix for Two Vectors}
\usage{
centered_kernel_matrix(
  first_vec_kernel,
  second_vec_kernel,
  centering_grid,
  hurst_coef
)
}
\arguments{
\item{first_vec_kernel}{Numeric vector. The first input vector.}

\item{second_vec_kernel}{Numeric vector. The second input vector.}

\item{centering_grid}{Numeric vector. The centering grid used for kernel centering.}

\item{hurst_coef}{Numeric. The Hurst coefficient for the Brownian kernel.}
}
\value{
A matrix of size \code{length(first_vec_kernel)} by \code{length(second_vec_kernel)},
where each element represents the centered kernel between the corresponding elements
of the first and second vectors.
}
\description{
This function calculates the centered kernel matrix between two input vectors,
based on the Brownian kernel with a specified Hurst coefficient. The resulting matrix
contains the centered kernel values for each pair of elements from the input vectors.
}
\examples{
vec1 <- c(1, 2, 7)
vec2 <- c(1, 2)
cent_grid <- sort(runif(1000, min = -1, max = 1))
h_result_matrix <- centered_kernel_matrix(vec1, vec2, cent_grid, hurst_coef = 0.5)
}
