% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_likelihood.R
\name{compute_marginal_likelihood_grid}
\alias{compute_marginal_likelihood_grid}
\title{Title: Compute Marginal Likelihood over a Grid of Parameters}
\usage{
compute_marginal_likelihood_grid(
  centered_kernel_mat_at_sampled,
  min_x,
  max_x,
  sampled_x,
  lambda_grid,
  tau_hat,
  initial_lambda_hat = 1,
  MC_iterations
)
}
\arguments{
\item{centered_kernel_mat_at_sampled}{A matrix of centered kernel values at sampled points.}

\item{lambda_grid}{A grid of lambda values to be evaluated.}

\item{MC_iterations}{Number of Monte Carlo iterations to perform.}

\item{centered_kernel_mat_at_grid}{A matrix of centered kernel values at grid points.}

\item{centered_kernel_self_grid}{A matrix of centered kernel self-values at grid points.}

\item{x_grid}{A grid of x values where the function is evaluated.}

\item{p_vec}{A vector of initial probabilities (default is a vector of 1's).}

\item{tau_grid}{A grid of tau values to be evaluated.}
}
\value{
A data frame containing lambda, tau, and their corresponding marginal log likelihoods.
}
\description{
This function computes the marginal likelihood for each combination of lambda and tau
in the provided grids.
}
