% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_densities.R
\name{densities_at_grid}
\alias{densities_at_grid}
\title{Compute Densities at Grid Points}
\usage{
densities_at_grid(
  centered_kernel_mat_at_grid,
  centered_kernel_self_grid,
  lambda_hat,
  weight_hat_vec
)
}
\arguments{
\item{centered_kernel_mat_at_grid}{A matrix (n x m) where n is the number of sampled points
and m is the number of grid points. The matrix should be a centered kernel matrix
evaluated at the grid points.}

\item{centered_kernel_self_grid}{A vector (length m) representing the diagonal of the centered kernel
matrix evaluated at the grid points.}

\item{lambda_hat}{A scalar value representing the estimated lambda.}

\item{weight_hat_vec}{A vector of weights (length n) corresponding to the sampled points.}
}
\value{
A vector of densities (length m) corresponding to the grid points.
}
\description{
This function computes the densities at grid points based on a centered kernel matrix,
estimated lambda, and a weight vector.
}
\examples{
# Example usage (assuming centered_kernel_mat_at_grid, centered_kernel_self_grid, lambda_hat, and weight_hat_vec are defined):
den_vec <- densities_at_grid(centered_kernel_mat_at_grid, centered_kernel_self_grid, lambda_hat, weight_hat_vec)
}
