% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dens_or_probs.R
\name{density_at_grid_R}
\alias{density_at_grid_R}
\title{Density at Grid Points}
\usage{
density_at_grid_R(
  centered_kernel_mat_at_grid,
  centerd_kernel_self_grid,
  lambda_hat,
  weight_hat_vec
)
}
\arguments{
\item{centered_kernel_mat_at_grid}{A matrix where each column represents the
centered kernel values at different grid points.}

\item{centerd_kernel_self_grid}{A vector containing the centered kernel values
for the grid points when compared to themselves.}

\item{lambda_hat}{A scalar parameter used to scale the contribution of the kernel
values and weights in the exponential function.}

\item{weight_hat_vec}{A vector of weights that are applied to the centered kernel
matrix in the computation.}
}
\value{
A vector containing the computed density values at each grid point.
}
\description{
This function calculates the density values at specific grid points using
a given centered kernel matrix, self-grid kernel values, a scaling parameter (lambda_hat),
and a vector of weights (weight_hat_vec).
}
\examples{
# Example usage of the function
density_at_grid(centered_kernel_mat_at_grid, centerd_kernel_self_grid, lambda_hat, weight_hat_vec)
}
