% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dens_or_probs.R
\name{density_at_sampled_x_R}
\alias{density_at_sampled_x_R}
\title{Density at Sampled Points}
\usage{
density_at_sampled_x_R(
  centered_kernel_mat_at_sampled,
  lambda_hat,
  weight_hat_vec
)
}
\arguments{
\item{centered_kernel_mat_at_sampled}{A matrix where each column represents the
centered kernel values at different sampled data points.}

\item{lambda_hat}{A scalar parameter used to scale the contribution of the kernel
values and weights in the exponential function.}

\item{weight_hat_vec}{A vector of weights that are applied to the centered kernel
matrix in the computation.}
}
\value{
A vector containing the computed density values at each sampled data point.
}
\description{
This function calculates the density values at sampled data points using
a given centered kernel matrix, a scaling parameter (lambda_hat), and a vector of weights (weight_hat_vec).
}
\examples{
# Example usage of the function
density_at_sampled_x(centered_kernel_mat_at_sampled, lambda_hat, weight_hat_vec)
}
