% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{format_number}
\alias{format_number}
\title{Format Number with Three Significant Digits}
\usage{
format_number(x)
}
\arguments{
\item{x}{A numeric value that you want to format.}
}
\value{
A character string representing the formatted number with three significant digits.
The output will be in scientific notation if the value of `x` is very small (less than 1e-3)
or very large (greater than 1e3).
}
\description{
This function formats a given numeric value `x` to three significant digits.
It uses scientific notation if the value is less than 1e-3 or greater than 1e3.
}
\examples{
format_number(3.456)
# Output: "3.46"

format_number(1.127e-12)
# Output: "1.13e-12"

format_number(123456)
# Output: "1.23e+05"

}
