% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dens_or_probs.R
\name{get_dens_or_prob}
\alias{get_dens_or_prob}
\title{Compute Densities for Sampled and Grid Points}
\usage{
get_dens_or_prob(
  centered_kernel_mat_at_sampled,
  centered_kernel_mat_at_grid,
  centered_kernel_self_grid,
  sampled_x,
  x_grid,
  lambda_hat,
  weight_hat_vec,
  type_of_p_is_prob = TRUE,
  type_of_q_is_prob = TRUE,
  method_of_p_calculation = "ordinary"
)
}
\arguments{
\item{centered_kernel_mat_at_sampled}{A square matrix (n x n) where n is the number of sampled points.
The matrix should be a centered kernel matrix evaluated at the sampled points.}

\item{centered_kernel_mat_at_grid}{A matrix (n x m) where n is the number of sampled points
and m is the number of grid points. The matrix should be a centered kernel matrix
evaluated at the grid points.}

\item{centered_kernel_self_grid}{A vector (length m) representing the diagonal of the centered kernel
matrix evaluated at the grid points.}

\item{sampled_x}{A vector of sampled points where the probabilities are evaluated.}

\item{x_grid}{A vector of grid points where the probabilities are evaluated.}

\item{lambda_hat}{A scalar value representing the estimated lambda.}

\item{weight_hat_vec}{A vector of weights (length n) corresponding to the sampled points.}

\item{type_of_p_is_prob}{A Boolean that specifies if the type of p is probability (TRUE) or it is density (FALSE).}

\item{type_of_q_is_prob}{A Boolean that specifies if the type of q is probability (TRUE) or it is density (FALSE).}

\item{method_of_p_calculation}{A string that shows by which approach we calculate the p. If it is ''ordinary'' it means that we used the ordinary approach of density at sampled points and if it is ''neighborhood_grid'' approach it means that we approximate sample densities or probabilities using grid densities.}
}
\value{
A list containing two elements:
        \item{sampled_x}{A vector of densities or normalized probabilities at the sampled points.}
        \item{grid_x}{A vector of densities or normalized probabilities at the grid points.}
}
\description{
This function computes densities for both sampled points and grid points.
The normalization is done by dividing the densities by the integral over the grid points.
}
\examples{
# Example usage (assuming inputs are defined):
probs <- get_dens_or_prob(centered_kernel_mat_at_sampled, centered_kernel_mat_at_grid,
                   centered_kernel_self_grid, sampled_x,
                   x_grid, lambda_hat, weight_hat_vec)
}
