% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dens_or_probs.R
\name{get_dens_wo_grid_with_base}
\alias{get_dens_wo_grid_with_base}
\title{Compute Densities for Sampled points without a grid}
\usage{
get_dens_wo_grid_with_base(
  centered_kernel_mat_at_sampled,
  min_x,
  max_x,
  sampled_x,
  lambda_hat,
  weight_hat_vec
)
}
\arguments{
\item{centered_kernel_mat_at_sampled}{A square matrix (n x n) where n is the number of sampled points.
The matrix should be a centered kernel matrix evaluated at the sampled points.}

\item{min_x}{A scalar value representing the minimum of x domain.}

\item{max_x}{A scalar value representing the maximum of x domain.}

\item{sampled_x}{A vector of sampled points where the probabilities are evaluated.}

\item{lambda_hat}{A scalar value representing the estimated lambda.}

\item{weight_hat_vec}{A vector of weights (length n) corresponding to the sampled points.}
}
\value{
A list containing an element:
        \item{sampled_x}{A vector of densities or normalized probabilities at the sampled points.}
}
\description{
This function computes densities for both sampled points and grid points.
The normalization is done by dividing the densities by the integral over the grid points.
}
\examples{
# Example usage (assuming inputs are defined):
probs <- get_dens_wo_grid(centered_kernel_mat_at_sampled,
                                  sampled_x,
                   x_grid, lambda_hat, weight_hat_vec)
}
