% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grid_approx_dens_or_probs.R
\name{get_grid_approx_dens_or_probs}
\alias{get_grid_approx_dens_or_probs}
\title{Get Grid Approximate Densities or Probabilities}
\usage{
get_grid_approx_dens_or_probs(sample, grid, dens_list)
}
\arguments{
\item{sample}{A numeric vector representing the sample points.}

\item{grid}{A numeric vector representing the grid.}

\item{dens_list}{A list containing the densities or probabilities at each grid point.}
}
\value{
A list containing two elements:
\item{prob}{A numeric vector of approximate probabilities for each sample point.}
\item{dens}{A numeric vector of approximate densities for each sample point.}
}
\description{
This function computes the approximate densities or probabilities for each
sample point based on a grid and a list of densities.
}
\examples{
get_grid_approx_dens_or_probs(c(1, 3, 5), c(0, 2, 4, 6), list(grid_x = c(0.1, 0.4, 0.3, 0.2)))
}
