% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{get_weights_wo_grid_BBsolve}
\alias{get_weights_wo_grid_BBsolve}
\title{Estimate Weights Using the Barzilai-Borwein Method (Optimized with Rcpp)}
\usage{
get_weights_wo_grid_BBsolve(
  lambda_hat,
  tau_hat,
  centered_kernel_mat_at_sampled,
  sampled_x,
  min_x,
  max_x,
  print_trace = FALSE
)
}
\arguments{
\item{lambda_hat}{A scalar representing the estimated lambda parameter.}

\item{tau_hat}{A scalar representing the estimated tau parameter.}

\item{centered_kernel_mat_at_sampled}{A matrix representing the centered kernel at sampled points.}

\item{sampled_x}{A vector of sampled points.}

\item{min_x}{The minimum domain value.}

\item{max_x}{The maximum domain value.}

\item{print_trace}{Logical; if TRUE, prints progress updates.}
}
\value{
A numeric vector of estimated weights.
}
\description{
This function estimates the weight vector using a Barzilai-Borwein method.
}
