% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gram_matrix.R
\name{gram_matrix}
\alias{gram_matrix}
\title{Compute Gram matrix for a list of vectors using a specified kernel function.}
\usage{
gram_matrix(
  vec_list,
  kernel_type = "rbf",
  kernel_params = list(length_scale = 1, degree = 2, free_add = 0, free_mult = 1,
    nu_matern = 1, centering_param = 5)
)
}
\arguments{
\item{vec_list}{A list of vectors for which the Gram matrix needs to be computed.}

\item{kernel_type}{The type of kernel function to be used (default: "rbf").}

\item{kernel_params}{A list of parameters for the kernel function (default values provided).}
}
\value{
A Gram matrix computed using the specified kernel function.
}
\description{
Compute Gram matrix for a list of vectors using a specified kernel function.
}
\examples{
vectors_list <- list(c(1, 2, 3), c(4, 5, 6), c(7, 8, 9))
result_gram_matrix <- gram_matrix(vectors_list, kernel_type = "linear")
print(result_gram_matrix)
}
