% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{initialize_excel_file}
\alias{initialize_excel_file}
\title{Initialize Excel File}
\usage{
initialize_excel_file(file_name)
}
\arguments{
\item{file_name}{A string specifying the name of the Excel file to be created.}
}
\value{
None
}
\description{
This function creates an Excel file with specified columns to store
simulation parameters and error rates.
}
\examples{
initialize_excel_file("simulation_results.xlsx")
}
