% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{is.singular}
\alias{is.singular}
\title{Check if a Matrix is Singular}
\usage{
is.singular(mat, tol = .Machine$double.eps)
}
\arguments{
\item{mat}{A numeric matrix to be checked for singularity.}

\item{tol}{A numeric value specifying the tolerance level for the determinant.
Default is `.Machine$double.eps`, the smallest positive floating-point number.}
}
\value{
A logical value: `TRUE` if the matrix is singular, `FALSE` otherwise.
}
\description{
This function checks if a given matrix is singular by calculating its determinant.
If the determinant is less than a specified tolerance, the matrix is considered singular.
}
\examples{
mat <- matrix(c(1, 2, 2, 4), nrow = 2)
is.singular(mat)  # Should return TRUE since the matrix is singular

}
