% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife_grid.R
\name{jackknife_weight_error_grid_nested_parallel}
\alias{jackknife_weight_error_grid_nested_parallel}
\title{Jackknife Weight Error Grid}
\usage{
jackknife_weight_error_grid_nested_parallel(
  centered_kernel_mat_at_sampled,
  centered_kernel_mat_at_grid,
  centered_kernel_self_grid,
  sampled_x,
  x_grid,
  lambda_hat_grid,
  tau_hat_grid,
  type_of_p_is_prob = TRUE,
  type_of_q_is_prob = TRUE,
  method_of_p_calculation = "ordinary",
  cloud_computing = FALSE,
  os = "Windows"
)
}
\arguments{
\item{centered_kernel_mat_at_sampled}{Centered kernel matrix at sampled points.}

\item{centered_kernel_mat_at_grid}{Centered kernel matrix at grid points.}

\item{centered_kernel_self_grid}{Centered kernel matrix for the grid points themselves.}

\item{sampled_x}{Sampled x values.}

\item{x_grid}{Grid of x values where the function is evaluated.}

\item{lambda_hat_grid}{A vector of lambda hat values.}

\item{tau_hat_grid}{A vector of tau hat values.}

\item{type_of_p_is_prob}{Logical. If TRUE, p is treated as probabilities.}

\item{type_of_q_is_prob}{Logical. If TRUE, q is treated as probabilities.}

\item{method_of_p_calculation}{Character. Method for p calculation.}
}
\value{
A data frame containing lambda_hat, tau_hat, and the corresponding Jackknife error (Jackknife_err).
}
\description{
Computes jackknife error for a grid of lambda_hat and tau_hat values.
}
\examples{
# Example usage:
# jackknife_weight_error_grid(...)
}
