% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{kernels}
\alias{kernels}
\alias{kernel}
\title{Kernel functions}
\usage{
kernel(
  x,
  fixed,
  type = "rbf",
  length_scale = 1,
  degree = 2,
  free_add = 0,
  free_mult = 1,
  nu_matern = 1,
  centering_param
)
}
\arguments{
\item{x}{Kernel free argument, the location to evaluate KDE (single scalar or vector)}

\item{fixed}{Kernel fixed argument (typically sample data vector or scalar)}

\item{type}{Kernel name (\code{default = "rbf"})}

\item{length_scale}{Kernel parameters, length scale}

\item{degree}{Kernel parameters, polynomial degree}

\item{free_add}{Kernel parameters, free additive parameter}

\item{free_mult}{Kernel parameters, free multiplicative parameter}

\item{nu_matern}{Kernel parameters, nu_matern for matern kernel

, polynomial degree, free parameters, etc. for the kernel (as half-width of the kernel) or \code{NULL}}
}
\description{
Functions for commonly used kernels for kernel density estimation. The density and cumulative distribution functions are provided.
}
\examples{
kernel(c(1,2,3),c(2,1,3), type = "linear")

}
\seealso{
\code{\link{norm}}, \code{\link{exp}}, \code{\link{sin}}, \code{\link{besselK}}
}
\author{
Pouya Roudaki \email{roudaki.pouya@gmail.com}.
}
\keyword{kernel}
