% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kme_shrinkage_loocv.R
\name{loocv_shr}
\alias{loocv_shr}
\title{Leave-One-Out Cross-Validation (LOOCV) error for shrinkage estimator}
\usage{
loocv_shr(gram, lambda, precomp_reg_inv = NULL)
}
\arguments{
\item{gram}{Matrix representing the Gram matrix.}

\item{lambda}{The shrinkage parameter.}
}
\value{
The LOOCV error estimate.
}
\description{
Leave-One-Out Cross-Validation (LOOCV) error for shrinkage estimator
}
