% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_likelihood.R
\name{marginal_likelihood}
\alias{marginal_likelihood}
\title{Title: Compute Marginal Likelihood}
\usage{
marginal_likelihood(
  centered_kernel_mat_at_sampled,
  centered_kernel_mat_at_grid,
  centered_kernel_self_grid,
  x_grid,
  p_old_vec = rep(1, nrow(centered_kernel_mat_at_sampled)),
  lambda_hat,
  tau_hat,
  MC_iterations
)
}
\arguments{
\item{centered_kernel_mat_at_sampled}{A matrix of centered kernel values at sampled points.}

\item{centered_kernel_mat_at_grid}{A matrix of centered kernel values at grid points.}

\item{centered_kernel_self_grid}{A matrix of centered kernel self-values at grid points.}

\item{x_grid}{A grid of x values where the function is evaluated.}

\item{p_old_vec}{A vector of initial probabilities (default is a vector of 1's).}

\item{lambda_hat}{A scalar parameter, lambda, to be used in the likelihood computation.}

\item{tau_hat}{A scalar parameter, tau, to be used in the likelihood computation.}

\item{MC_iterations}{Number of Monte Carlo iterations to perform.}
}
\value{
The log of the marginal likelihood.
}
\description{
This function calculates the marginal likelihood for a given set of parameters.
}
