% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_likelihood.R
\name{marginal_log_likelihood}
\alias{marginal_log_likelihood}
\title{Title: Compute Marginal Log Likelihood}
\usage{
marginal_log_likelihood(
  centered_kernel_mat_at_sampled,
  sampled_x,
  min_x,
  max_x,
  p_vec = rep(1, nrow(centered_kernel_mat_at_sampled)),
  lambda,
  tau,
  std_rnorm_matrix,
  MC_iterations
)
}
\arguments{
\item{centered_kernel_mat_at_sampled}{A matrix of centered kernel values at sampled points.}

\item{lambda}{A scalar parameter, lambda, to be used in the likelihood computation.}

\item{tau}{A scalar parameter, tau, to be used in the likelihood computation.}

\item{MC_iterations}{Number of Monte Carlo iterations to perform.}

\item{centered_kernel_mat_at_grid}{A matrix of centered kernel values at grid points.}

\item{centered_kernel_self_grid}{A matrix of centered kernel self-values at grid points.}

\item{x_grid}{A grid of x values where the function is evaluated.}

\item{dens_vec}{A vector of initial probabilities (default is a vector of 1's).}
}
\value{
The log of the marginal likelihood.
}
\description{
This function calculates the marginal likelihood for a given set of parameters.
}
