% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norms.R
\name{norm}
\alias{norm}
\title{Vector Norm Calculation}
\usage{
norm(x, p = 2)
}
\arguments{
\item{x}{Numeric or complex vector.}

\item{p}{Numeric value specifying the norm order (default is 2).}
}
\value{
The calculated vector norm.
}
\description{
This function calculates the vector norm of a numeric or complex vector.
}
\examples{
norm(c(3, 4)) # Calculates the Euclidean norm of the vector c(3, 4)
norm(c(1, 2, 3), p = 3) # Calculates the L3 norm of the vector c(1, 2, 3)

}
\references{
- Book: "Numerical Analysis" by Richard L. Burden and J. Douglas Faires.
}
\seealso{
\code{\link{abs}}, \code{\link{sum}}, \code{\link{max}}, \code{\link{min}}
}
\author{
Pouya Roudaki \email{roudaki.pouya@gmail.com}.
}
\keyword{norm,}
\keyword{vector}
