% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{ones}
\alias{ones}
\title{Create a Matrix of Ones}
\usage{
ones(n, m)
}
\arguments{
\item{n}{Integer. The number of rows in the matrix.}

\item{m}{Integer. The number of columns in the matrix.}
}
\value{
A matrix with `n` rows and `m` columns, where all elements are set to 1.
}
\description{
This function generates a matrix filled with ones, with dimensions specified by the user.
}
\examples{
# Create a 3x4 matrix of ones
ones(3, 4)
}
