% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kme_shrinkage_loocv.R
\name{p_i_lambda}
\alias{p_i_lambda}
\title{Computes the p_{i, lambda} vector for the i-th observation}
\usage{
p_i_lambda(gram, regularized_inv, i)
}
\arguments{
\item{gram}{Matrix representing the Gram matrix.}

\item{regularized_inv}{The regularized inverse of the Gram matrix.}

\item{i}{Index of the observation.}
}
\value{
A vector representing the p_{i, lambda}
}
\description{
Computes the p_{i, lambda} vector for the i-th observation
}
