% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R, R/kme_shrinkage.R
\name{rnorm_mixture_wrt_base_measure}
\alias{rnorm_mixture_wrt_base_measure}
\title{Generate Samples from a Normal Mixture Model with a Base Measure}
\usage{
rnorm_mixture_wrt_base_measure(
  n,
  probabilities,
  means,
  sds,
  lambda,
  centering_param
)

rnorm_mixture_wrt_base_measure(
  n,
  probabilities,
  means,
  sds,
  lambda,
  centering_param
)
}
\arguments{
\item{n}{Integer, number of samples to generate.}

\item{probabilities}{Numeric vector, probabilities for each component of the mixture.}

\item{means}{Numeric vector, means of each component of the mixture.}

\item{sds}{Numeric vector, standard deviations of each component of the mixture.}

\item{lambda}{Numeric, parameter for the base measure.}

\item{centering_param}{Numeric, centering parameter for the base measure.}
}
\value{
A numeric vector of generated samples.

A numeric vector of generated samples.
}
\description{
This function generates samples from a Gaussian mixture model with a base measure,
accounting for centering and lambda parameters.

This function generates samples from a Gaussian mixture model with a base measure,
accounting for centering and lambda parameters.
}
\examples{
samples <- rnorm_mixture_wrt_base_measure(1000, c(0.3, 0.7), c(-2, 2), c(1, 1.5), 0.5, 1)
samples <- rnorm_mixture_wrt_base_measure(1000, c(0.3, 0.7), c(-2, 2), c(1, 1.5), 0.5, 1)
}
