% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kme_shrinkage.R
\name{shr_est_KME}
\alias{shr_est_KME}
\title{Shrinkage Estimator of Kernel Mean Embedding(SKMSE)}
\usage{
shr_est_KME(
  evaluate_at,
  list_fixed,
  kernel_type = "rbf",
  kernel_params = list(length_scale = 1, degree = 2, free_add = 0, free_mult = 1,
    nu_matern = 1),
  lambda_tunner = 1,
  precomputed = list(gram = NULL, lambda = NULL, inverse_regularizer = NULL, beta_s =
    NULL)
)
}
\arguments{
\item{evaluate_at}{The point at which to evaluate the kernel mean embedding estimator function.}

\item{list_fixed}{A list of fixed points for the kernel mean embedding estimator function.}

\item{kernel_type}{Type of kernel function (default is "rbf" for Radial Basis Function).}

\item{kernel_params}{A list of kernel parameters, including length_scale, degree, free_add, free_mult, and nu_matern.}
}
\value{
The shrinkage estimator of Kernel Mean Embedding.
}
\description{
Compute the shrinkage estimator of Kernel Mean Embedding. More shrinkage in the direction of less variation
(More shrinkage in the direction of eigen vectors with lower eigen values)
}
\examples{
shr_est_KME(c(5, 3, 8), list(c(0, 1, 2), c(1, 2, 3)), kernel_type = "rbf", kernel_params = list(length_scale = 1, degree = 2, free_add = 0, free_mult = 1, nu_matern = 1))
}
\references{
Kernel Mean Shrinkage Estimators, Muandet et al.,
Journal of Machine Learning Research 17 (2016) 1-41 Submitted 5/14; Revised 9/15; Published 4/16,
Equation (32)
}
\author{
Pouya Roudaki \email{roudaki.pouya@gmail.com}.
}
