% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kme_standard.R
\name{std_est_KME}
\alias{std_est_KME}
\title{Standard Estimator of Kernel Mean Embedding(KMSE)}
\usage{
std_est_KME(
  evaluate_at,
  list_fixed,
  kernel_type = "rbf",
  kernel_params = list(length_scale = 1, degree = 2, free_add = 0, free_mult = 1,
    nu_matern = 1)
)
}
\arguments{
\item{evaluate_at}{The point at which to evaluate the kernel mean embedding estimator function.}

\item{list_fixed}{A list of fixed points for the kernel mean embedding estimator function.}

\item{kernel_type}{Type of kernel function (default is "rbf" for Radial Basis Function).}

\item{kernel_params}{A list of kernel parameters, including length_scale, degree, free_add, free_mult, and nu_matern.}
}
\value{
The standardized estimate using Kernel Mean Embedding.
}
\description{
Compute the standard estimator of Kernel Mean Embedding.
}
\examples{
std_est_KME(c(5, 3, 8), list(c(0, 1, 2), c(1, 2, 3)), kernel_type = "rbf", kernel_params = list(length_scale = 1, degree = 2, free_add = 0, free_mult = 1, nu_matern = 1))
}
\author{
Pouya Roudaki \email{roudaki.pouya@gmail.com}.
}
