% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_density_function.R
\name{true_density_function}
\alias{true_density_function}
\title{Compute True Density Based on Specified Distribution}
\usage{
true_density_function(grid, density_characterization)
}
\arguments{
\item{grid}{A numeric vector representing the grid points where the density should be computed.}

\item{density_characterization}{A list containing the density type and its parameters.}
}
\value{
A numeric vector containing the density values at the grid points.
}
\description{
Compute True Density Based on Specified Distribution
}
\examples{
grid <- seq(-5, 5, length.out = 100)

# Mixture Normal Example
params_mixture <- list(means = c(-2, 2), sds = c(1, 1.5), weights = c(0.4, 0.6))
density_characterization_mixture <- list(type = "mixture_normal", parameters = params_mixture)
true_density_mixture <- true_density_function(grid, density_characterization_mixture)
plot(grid, true_density_mixture, type = "l", main = "Mixture Normal Density", xlab = "x", ylab = "Density")

# Uniform Example
params_uniform <- list(min = -3, max = 3)
density_characterization_uniform <- list(type = "uniform", parameters = params_uniform)
true_density_uniform <- true_density_function(grid, density_characterization_uniform)
plot(grid, true_density_uniform, type = "l", main = "Uniform Density", xlab = "x", ylab = "Density")

# Poisson Example
grid_pos <- 0:10  # Poisson works with integer values
params_poisson <- list(lambda = 4)
density_characterization_poisson <- list(type = "poisson", parameters = params_poisson)
true_density_poisson <- true_density_function(grid_pos, density_characterization_poisson)
plot(grid_pos, true_density_poisson, type = "h", main = "Poisson Density", xlab = "x", ylab = "Density")

# Beta Example
grid_beta <- seq(0, 1, length.out = 100)  # Beta is defined on [0,1]
params_beta <- list(shape1 = 2, shape2 = 5)
density_characterization_beta <- list(type = "beta", parameters = params_beta)
true_density_beta <- true_density_function(grid_beta, density_characterization_beta)
plot(grid_beta, true_density_beta, type = "l", main = "Beta Density", xlab = "x", ylab = "Density")

}
