% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R, R/kme_shrinkage.R
\name{upper_tri_to_mat}
\alias{upper_tri_to_mat}
\title{Convert upper triangle vector to matrix}
\usage{
upper_tri_to_mat(upper_triangle_vector)

upper_tri_to_mat(upper_triangle_vector)
}
\arguments{
\item{upper_triangle_vector}{a vector including the upper triangle elements of
a matrix}
}
\value{
symmetric_matrix which is a symmetric matrix made by upper triangle vector

symmetric_matrix which is a symmetric matrix made by upper triangle vector
}
\description{
Convert upper triangle vector to matrix

Convert upper triangle vector to matrix
}
\author{
Pouya Roudaki \email{roudaki.pouya@gmail.com}.
}
