% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife_grid.R
\name{calc_jackknife_error}
\alias{calc_jackknife_error}
\title{Calculate Jackknife Error for a Given Pair of Parameters}
\usage{
calc_jackknife_error(
  lambda_hat,
  tau_hat,
  centered_kernel_mat_at_sampled,
  sampled_x,
  min_x,
  max_x
)
}
\arguments{
\item{lambda_hat}{Numeric. Regularization parameter for weight estimation.}

\item{tau_hat}{Numeric. Regularization parameter for controlling smoothness.}

\item{centered_kernel_mat_at_sampled}{A square matrix (n x n) representing the centered
kernel matrix evaluated at the sampled points, where n is the number of sampled points.}

\item{sampled_x}{A vector of sampled points for which the weights are to be estimated.}

\item{min_x}{A scalar representing the minimum value of the domain.}

\item{max_x}{A scalar representing the maximum value of the domain.}
}
\value{
Numeric. The calculated jackknife error for the given parameter pair.
}
\description{
Computes the jackknife error
given specific values of \code{lambda_hat} and \code{tau_hat}.
}
\examples{
# Example usage:
# calc_jackknife_error(0.1, 0.5, weight_hat, ...)
}
