% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weights.R
\name{get_weights_wo_grid_BBsolve}
\alias{get_weights_wo_grid_BBsolve}
\title{Estimate Weights Using the Newton-Raphson Method}
\usage{
get_weights_wo_grid_BBsolve(
  lambda_hat,
  tau_hat,
  centered_kernel_mat_at_sampled,
  sampled_x,
  min_x,
  max_x,
  print_trace = FALSE
)
}
\arguments{
\item{lambda_hat}{A scalar representing the estimated lambda parameter, which
controls the contribution of the kernel matrices to the weight estimation.}

\item{tau_hat}{A scalar representing the estimated tau parameter, which
determines the regularization strength applied to the weights.}

\item{centered_kernel_mat_at_sampled}{A square matrix (n x n) representing the centered
kernel matrix evaluated at the sampled points, where n is the number of sampled points.}

\item{sampled_x}{A vector of sampled points for which the weights are to be estimated.}

\item{min_x}{A scalar representing the minimum value of the domain.}

\item{max_x}{A scalar representing the maximum value of the domain.}

\item{print_trace}{Logical; if TRUE, prints progress updates during the Newton-Raphson iterations.}
}
\value{
A numeric vector of length n representing the estimated weight vector for the sampled points.
}
\description{
This function estimates the weight vector using an iterative Newton-Raphson method.
The method updates weights based on the provided kernel matrices, regularization
parameters, sampled points, and grid points.
}
