% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_likelihood_R.R
\name{optimize_marginal_log_likelihood}
\alias{optimize_marginal_log_likelihood}
\title{Optimize Marginal Log-Likelihood using L-BFGS-B Optimization}
\usage{
optimize_marginal_log_likelihood(
  centered_kernel_mat_at_sampled,
  min_x,
  max_x,
  sampled_x,
  initial_lambda = 1,
  initial_tau = 1,
  initial_w = rep(0, length(sampled_x)),
  MC_iterations,
  max.iterations = 1,
  censoring = FALSE
)
}
\arguments{
\item{centered_kernel_mat_at_sampled}{The kernel matrix centered at sampled points.}

\item{min_x}{Minimum value of the sampled domain.}

\item{max_x}{Maximum value of the sampled domain.}

\item{sampled_x}{Vector of sampled points.}

\item{initial_lambda}{Initial value for lambda (default: 1).}

\item{initial_tau}{Initial value for tau (default: 1).}

\item{initial_w}{Initial weights vector (default: zeros of sampled_x length).}

\item{MC_iterations}{Number of Monte Carlo iterations.}

\item{max.iterations}{Maximum number of iterations (default: 1).}

\item{censoring}{Boolean indicating whether censoring is applied (default: FALSE).}
}
\value{
A list containing:
	\item{lambda}{Optimized value of lambda.}
	\item{tau}{Optimized value of tau.}
	\item{max_marginal_log_likelihood}{Maximum marginal log-likelihood value.}
}
\description{
This function finds the optimal `lambda` and `tau` by maximizing the marginal log-likelihood
using an iterative approach. Instead of a grid search, it employs L-BFGS-B optimization
to efficiently search for the best parameters.
}
\examples{
\dontrun{
result <- optimize_marginal_log_likelihood(
  centered_kernel_mat_at_sampled, min_x, max_x, sampled_x,
  MC_iterations = 1000, max.iterations = 5
)
print(result)
}

}
